add log in and transaction

+++ database/migrations/2019_03_06_145324_create_farm_images_table.php
@@ -16,7 +16,7 @@ class CreateFarmImagesTable extends Migration
         Schema::dropIfExists('farm_images');
         Schema::create('farm_images', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->integer('farm_id')->unsigned();
+            $table->bigInteger('farm_id')->unsigned();
             $table->string('img_link');
             $table->timestamps();
         });
